DatabaseEntity moved from core.model -> model.core
