+++ src/main/java/egl/client/controller/info/EntityInfoController.java
@@ -1,7 +1,7 @@
 package egl.client.controller.info;
 
 import egl.client.controller.Controller;
-import egl.core.model.DatabaseEntity;
+import egl.client.model.core.DatabaseEntity;
 
 public interface EntityInfoController<T extends DatabaseEntity> extends Controller {
