+++ src/main/java/egl/client/controller/info/EntityInfoDialog.java
@@ -1,6 +1,6 @@
 package egl.client.controller.info;
 
-import egl.core.model.DatabaseEntity;
+import egl.client.model.core.DatabaseEntity;
 import javafx.scene.control.Alert;
 import javafx.scene.control.ButtonType;
 import javafx.scene.control.Dialog;
