+++ src/main/java/egl/client/repository/EntityRepository.java
@@ -1,6 +1,6 @@
 package egl.client.repository;
 
-import egl.core.model.DatabaseEntity;
+import egl.client.model.core.DatabaseEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EntityRepository<T extends DatabaseEntity> extends JpaRepository<T, Long> { }
