+++ src/main/java/egl/core/model/statistic/TaskStatistic.java
@@ -1,6 +1,6 @@
 package egl.core.model.statistic;
 
-import egl.core.model.DatabaseEntity;
+import egl.client.model.core.DatabaseEntity;
 import egl.core.model.task.Task;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
